+++ entra-id-dart-auth-sdk/lib/src/api_clients/entra_id_http_client.dart
@@ -356,6 +356,8 @@ class EntraIdHttpClient {
       }
 
       return _handleResponse(uri, response);
+    } on HttpException {
+      rethrow;
     } catch (e) {
       // Network errors / unreachable server
       throw HttpException(